:technologist: (smtp) Use secure prop instead of ignoreTLS
BREAKING CHANGE: The SMTP_IGNORE_TLS environment has been removed in favore of SMTP_SECURE. If
SMTP_SECURE is true,  the connection will use TLS when connecting to server. If false (the default)
then TLS is used if server supports the STARTTLS extension. In most cases set this value to true if
you are connecting to port 465. For port 587 or 25 keep it false